+++ README.md
@@ -62,7 +62,8 @@ page.
 
 ### Available methods
 See : 
-- [ITwitterClient.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/ITwitterClient.java)
+- [ITwitterClientV2.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/ITwitterClientV2.java)
+- [ITwitterClientV1.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/ITwitterClientV1.java)
 - [Tweet.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/dto/tweet/Tweet.java)
 - [User.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/dto/user/User.java)
 
@@ -73,7 +74,7 @@ See :
 ### Contribution
 If you want to contribute to the project, don't hesitate to submit pull requests.
 To add a new feature :
-- Create the interface method in the related interface (i.e [ITwitterClient.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/ITwitterClient.java))
+- Create the interface method in the related interface (i.e [ITwitterClientV2.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/ITwitterClientV2.java))
 - If needed, add the endpoint URL in [URLHelper.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/helpers/URLHelper.java)
 - Implement your method in the child class (i.e [TwitterClient.java](https://github.com/redouane59/twitter-client/blob/master/src/main/java/com/github/redouane59/twitter/TwitterClient.java))
 - Don't forget to add your unit tests in `src/test/java/com/github/redouane59/twitter/unit`
