+++ src/main/java/com/github/redouane59/twitter/ITwitterClientV1.java
@@ -3,18 +3,12 @@ package com.github.redouane59.twitter;
 import com.github.redouane59.RelationType;
 import com.github.redouane59.twitter.dto.others.RateLimitStatus;
 import com.github.redouane59.twitter.dto.others.RequestToken;
-import com.github.redouane59.twitter.dto.stream.StreamRules.StreamMeta;
-import com.github.redouane59.twitter.dto.stream.StreamRules.StreamRule;
 import com.github.redouane59.twitter.dto.tweet.Tweet;
-import com.github.redouane59.twitter.dto.tweet.TweetV1;
 import com.github.redouane59.twitter.dto.user.User;
-import java.io.File;
-import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.function.Consumer;
 
-public interface ITwitterClient {
+public interface ITwitterClientV1 {
 
   /**
    * Get a list of the user followers calling https://api.twitter.com/1.1/followers
@@ -99,38 +93,6 @@ public interface ITwitterClient {
   RelationType getRelationType(String userId1, String userId2);
 
   /**
-   * Retreive a user from his screen name calling https://api.twitter.com/2/users/
-   *
-   * @param userName the name of the targeted user
-   * @return an user object related to the targeted user
-   */
-  User getUserFromUserName(String userName);
-
-  /**
-   * Retreive a user from his id calling https://api.twitter.com/2/users/
-   *
-   * @param userId the id of the user
-   * @return an user object related to the targeted user
-   */
-  User getUserFromUserId(String userId);
-
-  /**
-   * Retreive a list of users from their usernames calling https://api.twitter.com/2/users/
-   *
-   * @param userNames the names of the targeted user
-   * @return an list of user objects related to the targeted users
-   */
-  List<User> getUsersFromUserNames(List<String> userNames);
-
-  /**
-   * Retreive a list of users from their ids calling https://api.twitter.com/2/users/
-   *
-   * @param userIds the id of the user
-   * @return an list of user object related to the targeted users
-   */
-  List<User> getUsersFromUserIds(List<String> userIds);
-
-  /**
    * Follow a user calling https://api.twitter.com/1.1/friendships/
    *
    * @param userId the id of the user
@@ -171,22 +133,6 @@ public interface ITwitterClient {
   List<Tweet> getUserLastTweets(String userId, int count);
 
   /**
-   * Get a tweet from its id calling https://api.twitter.com/2/tweets
-   *
-   * @param tweetId id of the tweet
-   * @return a tweet object
-   */
-  Tweet getTweet(String tweetId);
-
-  /**
-   * Get a tweet list from their id calling https://api.twitter.com/2/tweets
-   *
-   * @param tweetIds the ids of the tweets
-   * @return a tweet object list
-   */
-  List<Tweet> getTweets(List<String> tweetIds);
-
-  /**
    * Get the most recent Tweets liked calling https://api.twitter.com/1.1/favorites/list.json
    *
    * @param userId id of the user
@@ -194,33 +140,6 @@ public interface ITwitterClient {
    * @return a list of liked tweets
    */
   List<Tweet> getFavorites(String userId, int count);
-
-  /**
-   * Hide/Unide a reply using https://api.twitter.com/labs/2/tweets/:id/hidden
-   *
-   * @param tweetId id of the concerned reply
-   * @param hide true to hide the reply, false to unide it
-   * @return the hidden state
-   */
-  boolean hideReply(String tweetId, boolean hide);
-
-  /**
-   * Search tweets from last 7 days calling https://api.twitter.com/2/tweets/search
-   *
-   * @param query the search query
-   * @param fromDate the start date
-   * @param toDate the end date
-   * @return a list of tweets
-   */
-  List<Tweet> searchForTweetsWithin7days(String query, LocalDateTime fromDate, LocalDateTime toDate);
-
-  /**
-   * Search tweets from last 7 days calling https://api.twitter.com/2/tweets/search
-   *
-   * @param query the search query
-   * @return a list of tweets
-   */
-  List<Tweet> searchForTweetsWithin7days(String query);
 
   /**
    * Search tweets from last 30 days calling https://api.twitter.com/1.1/tweets/search/30day/dev.json Your development environment name should be
@@ -243,52 +162,7 @@ public interface ITwitterClient {
    * @return a list of tweets
    */
   List<Tweet> searchForTweetsArchive(String query, LocalDateTime fromDate, LocalDateTime toDate);
-
-  /**
-   * Stream using previous set up filters calling https://api.twitter.com/2/tweets/search/stream
-   */
-  void startFilteredStream(Consumer<Tweet> consumer);
-
-  /**
-   * add a filtered stream rule calling https://api.twitter.com/2/tweets/search/stream/rules
-   *
-   * @param value the rule value
-   * @param tag the rule associated tag
-   * @return the created rules
-   */
-  StreamRule addFilteredStreamRule(String value, String tag);
-
-  /**
-   * Delete a filtered stream rule calling https://api.twitter.com/2/tweets/search/stream/rules
-   *
-   * @param ruleValue the value of the rule to delete
-   * @return a StreamMeta object resuming the operation
-   */
-  StreamMeta deleteFilteredStreamRule(String ruleValue);
-
-  /**
-   * Retrieve the filtered stream rules calling https://api.twitter.com/2/tweets/search/stream/rules
-   *
-   * @return a filtered stream rules list
-   */
-  List<StreamRule> retrieveFilteredStreamRules();
-
-  /**
-   * Parse the Twitter extract data archive file in a List of Tweets
-   *
-   * @param file the downloaded file on https://twitter.com/settings/your_twitter_data converted in .json format
-   * @return the list of tweets
-   * @throws IOException if excpetion when reading file
-   */
-  List<TweetV1> readTwitterDataFile(File file) throws IOException;
-
-  /**
-   * Get a bearer token (oAuth2) calling https://api.twitter.com/oauth2/token
-   *
-   * @return the generated token
-   */
-  String getBearerToken();
-
+  
   /**
    * Get token and secret token (oAuth1) calling https://api.twitter.com/oauth/request_token
    *
