+++ src/test/java/com/github/redouane59/twitter/nrt/ITwitterClientV2Test.java
@@ -2,13 +2,10 @@ package com.github.redouane59.twitter.nrt;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.github.redouane59.RelationType;
 import com.github.redouane59.twitter.TwitterClient;
-import com.github.redouane59.twitter.dto.others.RequestToken;
 import com.github.redouane59.twitter.dto.stream.StreamRules.StreamMeta;
 import com.github.redouane59.twitter.dto.stream.StreamRules.StreamRule;
 import com.github.redouane59.twitter.dto.tweet.Tweet;
@@ -18,7 +15,6 @@ import com.github.redouane59.twitter.dto.user.User;
 import com.github.redouane59.twitter.signature.TwitterCredentials;
 import java.io.File;
 import java.io.IOException;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
@@ -26,7 +22,7 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 @Disabled
-public class TwitterClientTest {
+public class ITwitterClientV2Test {
 
   private static TwitterClient twitterClient;
 
@@ -35,32 +31,6 @@ public class TwitterClientTest {
     String credentialPath = "C:/Users/Perso/Documents/GitHub/twitter-credentials.json";
     twitterClient = new TwitterClient(TwitterClient.OBJECT_MAPPER
                                           .readValue(new File(credentialPath), TwitterCredentials.class));
-  }
-
-  @Test
-  public void testGetFollowingIdsById() {
-    List<String> followings = twitterClient.getFollowingIds("882266619115864066");
-    assertTrue(followings.size() > 200);
-  }
-
-  @Test
-  public void testGetFollowersIdsById() {
-    List<String> followers = twitterClient.getFollowerIds("882266619115864066");
-    assertTrue(followers.size() > 200);
-  }
-
-  @Test
-  public void testGetFollowersUsersById() {
-    List<User> followers = twitterClient.getFollowerUsers("882266619115864066");
-    assertTrue(followers.size() > 200);
-  }
-
-  @Test
-  public void testFriendshipByIdYes() {
-    String       userId1 = "92073489";
-    String       userId2 = "723996356";
-    RelationType result  = twitterClient.getRelationType(userId1, userId2);
-    assertEquals(RelationType.FRIENDS, result);
   }
 
   @Test
@@ -88,14 +58,6 @@ public class TwitterClientTest {
   }
 
   @Test
-  public void testFriendshipByIdNo() {
-    String       userId1 = "92073489";
-    String       userId2 = "1976143068";
-    RelationType result  = twitterClient.getRelationType(userId1, userId2);
-    assertNotEquals(RelationType.FRIENDS, result);
-  }
-
-  @Test
   public void testGetUserInfoName() {
     String userId = "92073489";
     User   user   = twitterClient.getUserFromUserId(userId);
@@ -120,7 +82,6 @@ public class TwitterClientTest {
   public void testGetUserInfoStatusesCount() {
     String userId = "92073489";
     User   user   = twitterClient.getUserFromUserId(userId);
-
     assertTrue(user.getTweetCount() > 0);
   }
 
@@ -143,64 +104,6 @@ public class TwitterClientTest {
     assertEquals("Soltana", result.get(1).getName());
   }
 
-  @Test
-  public void testGetRateLimitStatus() {
-    assertNotEquals(null, twitterClient.getRateLimitStatus());
-  }
-
-  @Test
-  public void testRelationBetweenUsersIdFriends() {
-    String       userId1 = "92073489";
-    String       userId2 = "723996356";
-    RelationType result  = twitterClient.getRelationType(userId1, userId2);
-    assertEquals(RelationType.FRIENDS, result);
-  }
-
-  @Test
-  public void testRelationBetweenUsersIdNone() {
-    String       userId1 = "92073489";
-    String       userId2 = "1976143068";
-    RelationType result  = twitterClient.getRelationType(userId1, userId2);
-    assertEquals(RelationType.NONE, result);
-  }
-
-  @Test
-  public void testRelationBetweenUsersIdFollowing() {
-    String       userId1 = "92073489";
-    String       userId2 = "126267113";
-    RelationType result  = twitterClient.getRelationType(userId1, userId2);
-    assertEquals(RelationType.FOLLOWING, result);
-  }
-
-  @Test
-  public void testRelationBetweenUsersIdFollower() {
-    String       userId1 = "92073489";
-    String       userId2 = "1218125226095054848";
-    RelationType result  = twitterClient.getRelationType(userId1, userId2);
-    assertEquals(RelationType.FOLLOWER, result);
-  }
-
-  @Test
-  public void testGetRetweetersId() {
-    String tweetId = "1078358350000205824";
-    assertTrue(twitterClient.getRetweetersId(tweetId).size() > 10);
-  }
-
-  @Test
-  public void testGetLastTweetByUserName() {
-    String      userName = "RedTheOne";
-    List<Tweet> response = twitterClient.getUserLastTweets(userName, 2);
-    assertTrue(response.get(0).getLang().equals("fr")
-               || response.get(1).getLang().equals("fr"));
-  }
-
-  @Test
-  public void testGetLastTweetByUserId() {
-    String      userId   = "92073489";
-    List<Tweet> response = twitterClient.getUserLastTweets(userId, 3);
-    assertTrue(response.get(0).getLang().equals("fr")
-               || response.get(1).getLang().equals("fr"));
-  }
 
   @Test
   public void testGetTweetById() {
@@ -220,15 +123,6 @@ public class TwitterClientTest {
   }
 
   @Test
-  public void testGetOauth1Token() {
-    TwitterClient.TWITTER_CREDENTIALS.setAccessToken("");
-    TwitterClient.TWITTER_CREDENTIALS.setAccessTokenSecret("");
-    RequestToken result = twitterClient.getOauth1Token();
-    assertTrue(result.getOauthToken().length() > 1);
-    assertTrue(result.getOauthTokenSecret().length() > 1);
-  }
-
-  @Test
   @Disabled
   public void testGetTweetDataFile() throws IOException {
     File          file   = new File(this.getClass().getClassLoader().getResource("tweet.json").getFile());
@@ -238,33 +132,6 @@ public class TwitterClientTest {
     assertNotNull(result.get(0).getId());
     assertNotNull(result.get(0).getText());
     assertNotNull(result.get(0).getInReplyToUserId());
-  }
-
-  @Test
-  public void testFollowAndUnfollow() {
-    User user         = twitterClient.getUserFromUserName("red1");
-    User followedUser = twitterClient.follow(user.getId());
-    assertEquals("red1", followedUser.getName());
-    User unfollowedUser = twitterClient.unfollow(user.getId());
-    assertEquals("red1", unfollowedUser.getName());
-    assertEquals(RelationType.NONE, twitterClient.getRelationType("92073489", "66533"));
-  }
-
-  @Test
-  public void testLikeTweet() {
-    Tweet likedTweet = twitterClient.likeTweet("1107533");
-    assertEquals("1107533", likedTweet.getId());
-    Tweet unlikedTweet = twitterClient.unlikeTweet("1107533");
-    assertEquals("1107533", unlikedTweet.getId());
-  }
-
-  @Test
-  public void testPostTweet() {
-    String text   = "API Test " + LocalDateTime.now() + " #TwitterAPI";
-    Tweet  result = twitterClient.postTweet(text);
-    assertNotNull(result);
-    assertNotNull(result.getId());
-    assertEquals(text, result.getText());
   }
 
   @Test
@@ -280,14 +147,6 @@ public class TwitterClientTest {
     assertTrue(tweet.getReplyCount() >= 0);
     assertTrue(tweet.getLikeCount() >= 0);
     assertNotNull(tweet.getLang());
-  }
-
-  @Test
-  public void testGetFavorites() {
-    int         count     = 1500;
-    List<Tweet> favorites = twitterClient.getFavorites("92073489", count);
-    assertNotNull(favorites);
-    assertTrue(favorites.size() > count);
   }
 
   @Test
@@ -329,23 +188,6 @@ public class TwitterClientTest {
     StreamMeta streamMeta = twitterClient.deleteFilteredStreamRule(ruleName);
     assertNotNull(streamMeta);
   }
-
-    /*
-    @Test
-    public void testSearchTweets30days(){
-        LocalDateTime startDate = DateUtils.truncate(new Date(),Calendar.MONTH);
-        LocalDateTime endDate = DateUtils.addDays(startDate, 1);
-        List<ITweet> result = twitterClient.searchForTweetsWithin30days("@RedTheOne -RT",startDate, endDate);
-        assertTrue(result.size()>0);
-    }
-
-    @Test
-    public void testSearchTweetsArchive(){
-        LocalDateTime startDate = DateUtils.truncate(ConverterHelper.dayBeforeNow(60),Calendar.MONTH);
-        LocalDateTime endDate = DateUtils.addDays(startDate, 1);
-        List<ITweet> result = twitterClient.searchForTweetsArchive("@RedTheOne -RT",startDate, endDate);
-        assertTrue(result.size()>0);
-    } */
 
   @Test
   public void testGetBearerToken() {
